+++ source/ada/lsp-ada_handlers.adb
@@ -18,6 +18,7 @@
 with Ada.Characters.Latin_1;
 with Ada.Strings.Unbounded;
 with Ada.Strings.UTF_Encoding;
+with Ada.Tags;
 with Ada.Unchecked_Deallocation;
 
 with GNAT.OS_Lib;
@@ -1963,8 +1964,7 @@ package body LSP.Ada_Handlers is
       --  Handle the case where we're loading the implicit project: do
       --  we need to add the directory in which the document is open?
 
-      if LSP.Ada_Project_Loading.Is_Implicit_Fallback (Self.Project_Status)
-      then
+      if Self.Project_Status.Is_Implicit_Fallback then
          declare
             Dir : constant GNATCOLL.VFS.Virtual_File := Self.To_File (URI).Dir;
          begin
@@ -2497,6 +2497,28 @@ package body LSP.Ada_Handlers is
              (Self'Unchecked_Access);
       end if;
 
+      --  If settings were given in initializationOptions, parse and apply them.
+      if not Value.initializationOptions.Is_Empty then
+         Self.Tracer.Trace ("Processing initializationOptions from initialize request");
+         declare
+            Reload : Boolean;
+         begin
+            --  The expected structure is this:
+            --     "initializationOptions": {
+            --        "ada": {
+            --           "projectFile": "...",
+            --           "scenarioVariables": ...,
+            --           ...
+            --        }
+            --     }
+            Self.Configuration.Read_JSON (Value.initializationOptions, Reload);
+
+            --  We don't load the project here because we can't send progress
+            --  notifications to the client before receiving the 'initialized'
+            --  notification. See On_Initialized_Notification.
+         end;
+      end if;
+
       Self.Sender.On_Initialize_Response (Id, Response);
 
       Log_Info.a_type := LSP.Enumerations.Log;
@@ -2504,6 +2526,42 @@ package body LSP.Ada_Handlers is
       Log_Info.message.Append (Self.Tracer.Location);
       Self.Sender.On_LogMessage_Notification (Log_Info);
    end On_Initialize_Request;
+
+   ---------------------------------
+   -- On_Initialized_Notification --
+   ---------------------------------
+
+   overriding procedure On_Initialized_Notification
+     (Self  : in out Message_Handler;
+      Value : LSP.Structures.InitializedParams) is
+   begin
+      --  The client is notifying us that it has initialized. If a project was
+      --  provided with the initialize request, we can load it and start
+      --  communicating with the client.
+      --
+      --  We only perform loading in case ada.projectFile was provided with the
+      --  initialize request. That's because some clients don't provide
+      --  settings in the initialize request, and instead send a
+      --  onDidChangeConfiguration notification immediately after
+      --  initialization. In this scenario, loading a project unconditionally
+      --  here would result in automatically searching for a project (unique
+      --  project at the root, or implicit fallback) and loading it, only to
+      --  load another project shortly after, upon receiving
+      --  onDidChangeConfiguration. To avoid the first useless load, we only
+      --  load a project here if it was specified in the initialize request.
+      --
+      --  Moreover, there is an impact on legacy tests. The prior project
+      --  loading policy was to wait for the first onDidChangeConfiguration
+      --  notification to obtain settings and load a project. Many existing
+      --  tests do not set the project in the initialize request and don't
+      --  expect messages pertaining to project loading after the initialize
+      --  request.  If we were to always load the project here, tests would
+      --  receive different messages and potentially fail. So the conditional
+      --  project loading here is backwards compatible with existing tests.
+      if not Self.Configuration.Project_File.Is_Empty then
+         LSP.Ada_Handlers.Project_Loading.Ensure_Project_Loaded (Self);
+      end if;
+   end On_Initialized_Notification;
 
    ---------------------------------
    -- On_OnTypeFormatting_Request --
@@ -3100,8 +3158,35 @@ package body LSP.Ada_Handlers is
    begin
       Value.Visit_Server_Receiver (Self);
    exception
+      when E : Libadalang.Common.Property_Error =>
+         --  Many LAL queries can raise Property_Error because sources can be
+         --  inconsistent while they're being edited. We don't want those
+         --  errors to be visible because they would create a lot of noise in
+         --  the UX. But we still want to get a chance to investigate them when
+         --  they indicate real problems, so we log them to the trace file.
+         Self.Tracer.Trace_Exception
+           (E,
+            VSS.Strings.Conversions.To_Virtual_String
+              ("Exception while handling " & Ada.Tags.Expanded_Name (Value'Tag)));
+
       when E : others =>
-         Self.Tracer.Trace_Exception (E, "On_Server_Notification");
+         --  Errors other than Property_Error indicate real problems to
+         --  investigate. But in the current state of the project these occur
+         --  often and can create a lot of noise to the User.
+         --
+         --  So we choose not to report them to the User and only log them in
+         --  the trace file for investigation when the problem is visible to
+         --  the User by other means.
+         --
+         --  TODO consider sending errors to the client in testing campaigns to
+         --  capture more issues
+         --  TODO when the ALS is more stable and fewer errors occur, consider
+         --  sending these errors through the LSP window/logMessage
+         --  notification
+         Self.Tracer.Trace_Exception
+           (E,
+            VSS.Strings.Conversions.To_Virtual_String
+              ("Exception while handling " & Ada.Tags.Expanded_Name (Value'Tag)));
    end On_Server_Notification;
 
    -----------------------
@@ -3136,7 +3221,18 @@ package body LSP.Ada_Handlers is
       end if;
 
    exception
-      when Libadalang.Common.Property_Error =>
+      when E : Libadalang.Common.Property_Error =>
+         --  Many LAL queries can raise Property_Error because sources can be
+         --  inconsistent while they're being edited. We don't want those
+         --  errors to be visible because they would create a lot of noise in
+         --  the UX. But we still want to get a chance to investigate them when
+         --  they indicate real problems, so we log them to the trace file.
+         Self.Tracer.Trace_Exception
+           (E,
+            VSS.Strings.Conversions.To_Virtual_String
+              ("Exception while handling " & Ada.Tags.Expanded_Name (Value'Tag)));
+
+         --  Send an empty response to the client to mask the error.
          declare
             R : LSP.Ada_Empty_Handlers.Empty_Message_Handler (Self.Sender);
          begin
@@ -3145,13 +3241,18 @@ package body LSP.Ada_Handlers is
 
       when E : others =>
          declare
+            Msg_Prefix : constant String := "Exception while handling "
+               & Ada.Tags.Expanded_Name (Value'Tag);
             Message : constant VSS.Strings.Virtual_String :=
               VSS.Strings.Conversions.To_Virtual_String
-                ("Exception: " & Ada.Exceptions.Exception_Information (E));
+                (Msg_Prefix & Ada.Characters.Latin_1.LF
+                 & Ada.Exceptions.Exception_Information (E));
 
          begin
-            Self.Tracer.Trace_Exception (E, "On_Server_Request");
+            Self.Tracer.Trace_Exception
+              (E, VSS.Strings.Conversions.To_Virtual_String (Msg_Prefix));
 
+            --  Send an error response to the client.
             Self.Sender.On_Error_Response
               (Value.Id,
                (code    => LSP.Enumerations.InternalError,